Compiler warning
Source commit: 2b3affc7251916fc2182957560f7f02bcb639135

+++ libs/libs3/src/request.c
@@ -1316,7 +1316,7 @@ static S3Status setup_neon(Request *request,
         {
             *colon = '\0';
         }
-        for (int i = 0; i < strlen(params->bucketContext.bucketName); i++)
+        for (size_t i = 0; i < strlen(params->bucketContext.bucketName); i++)
         {
             if (hostName[i] == '.')
             {
